+++ ClientSave
@@ -1,0 +1,2 @@
+/home/matjaz/Programming/File-transfer/Data/Folder
+/home/matjaz/Programming/File-transfer/Data/Text File.txt
