+++ ServerSave
@@ -1,0 +1,4 @@
+/home/matjaz/Programming/File-transfer/Data/Folder
+/home/matjaz/Programming/File-transfer/Data/Text File.txt
+ile.txt
+/home/matjaz/Programming/File-transfer/Received/Folder
